+++ src/utils/dateInfo.js
@@ -8,6 +8,7 @@ import {
 } from './typeCheckers';
 import defaults from './defaults';
 import { mixinOptionalProps, getMonthComps } from './helpers';
+import { parse } from './fecha';
 
 const millisecondsPerDay = 24 * 60 * 60 * 1000;
 
@@ -25,9 +26,11 @@ function findShallowIntersectingRange(date1, date2) {
   if (thisRange.start) {
     // Use this definite start date if other start date is infinite
     if (!otherRange.start) start = thisRange.start;
-    // Otherwise, use the earliest start date
-    else start = thisRange.start < otherRange.start ? thisRange.start : otherRange.start;
-  // Other start date exists
+    else
+      // Otherwise, use the earliest start date
+      start =
+        thisRange.start < otherRange.start ? thisRange.start : otherRange.start;
+    // Other start date exists
   } else if (otherRange.start) {
     // Use other definite start date as this one is infinite
     start = otherRange.start;
@@ -52,7 +55,9 @@ function dateShallowIncludesDate(date1, date2) {
   if (date1.isDate) {
     if (date2.isDate) return date1.dateTime === date2.dateTime;
     if (!date2.startTime || !date2.endTime) return false;
-    return date1.dateTime === date2.startTime && date1.dateTime === date2.endTime;
+    return (
+      date1.dateTime === date2.startTime && date1.dateTime === date2.endTime
+    );
   }
   // Second date is simple date and first is date range
   if (date2.isDate) {
@@ -69,7 +74,10 @@ function dateShallowIncludesDate(date1, date2) {
 // Determines if first date partially intersects second date
 // NOTE: This is a shallow test (no patterns tested)
 function dateShallowIntersectsDate(date1, date2) {
-  if (date1.isDate) return date2.isDate ? date1.dateTime === date2.dateTime : dateShallowIncludesDate(date2, date1);
+  if (date1.isDate)
+    return date2.isDate
+      ? date1.dateTime === date2.dateTime
+      : dateShallowIncludesDate(date2, date1);
   if (date2.isDate) return dateShallowIncludesDate(date1, date2);
   // Both ranges
   if (date1.start && date2.end && date1.start > date2.end) return false;
@@ -83,12 +91,14 @@ export function getDayFromDate(date) {
   const year = date.getUTCFullYear();
   const comps = getMonthComps(month, year);
   const day = date.getDate();
-  const dayFromEnd = (comps.days - day) + 1;
+  const dayFromEnd = comps.days - day + 1;
   const weekday = date.getDay() + 1;
-  const weekdayOrdinal = Math.floor(((day - 1) / 7) + 1);
-  const weekdayOrdinalFromEnd = Math.floor(((comps.days - day) / 7) + 1);
-  const week = Math.ceil((day + Math.abs(comps.firstWeekday - comps.firstDayOfWeek)) / 7);
-  const weekFromEnd = (comps.weeks - week) + 1;
+  const weekdayOrdinal = Math.floor((day - 1) / 7 + 1);
+  const weekdayOrdinalFromEnd = Math.floor((comps.days - day) / 7 + 1);
+  const week = Math.ceil(
+    (day + Math.abs(comps.firstWeekday - comps.firstDayOfWeek)) / 7,
+  );
+  const weekFromEnd = comps.weeks - week + 1;
   return {
     day,
     dayFromEnd,
@@ -130,7 +140,10 @@ function iterateDatesInRange({ start, end }, func) {
 function startOfWeek(date) {
   const day = date.getDay() + 1;
   const { firstDayOfWeek } = defaults;
-  const daysToAdd = day >= firstDayOfWeek ? firstDayOfWeek - day : -(7 - (firstDayOfWeek - day));
+  const daysToAdd =
+    day >= firstDayOfWeek
+      ? firstDayOfWeek - day
+      : -(7 - (firstDayOfWeek - day));
   return addDays(date, daysToAdd);
 }
 
@@ -147,45 +160,52 @@ function diffInYears(d1, d2) {
 }
 
 function diffInMonths(d1, d2) {
-  return (diffInYears(d1, d2) * 12) + (d2.getMonth() - d1.getMonth());
+  return diffInYears(d1, d2) * 12 + (d2.getMonth() - d1.getMonth());
 }
 
 const _patterns = {
   dailyInterval: {
-    test: (day, interval, { start }) => diffInDays(start || new Date(), day.date) % interval === 0,
+    test: (day, interval, { start }) =>
+      diffInDays(start || new Date(), day.date) % interval === 0,
   },
   weeklyInterval: {
-    test: (day, interval, { start }) => diffInWeeks(start || new Date(), day.date) % interval === 0,
+    test: (day, interval, { start }) =>
+      diffInWeeks(start || new Date(), day.date) % interval === 0,
   },
   monthlyInterval: {
-    test: (day, interval, { start }) => diffInMonths(start || new Date(), day.date) % interval === 0,
+    test: (day, interval, { start }) =>
+      diffInMonths(start || new Date(), day.date) % interval === 0,
   },
   yearlyInterval: {
-    test: () => (day, interval, { start }) => diffInYears(start || new Date(), day.date) % interval === 0,
+    test: () => (day, interval, { start }) =>
+      diffInYears(start || new Date(), day.date) % interval === 0,
   },
   days: {
     validate: days => (isArray(days) ? days : [parseInt(days, 10)]),
-    test: (day, days) => days.includes(day.day) || days.includes(-day.dayFromEnd),
+    test: (day, days) =>
+      days.includes(day.day) || days.includes(-day.dayFromEnd),
   },
   weekdays: {
-    validate: weekdays => (isArray(weekdays) ? weekdays : [parseInt(weekdays, 10)]),
+    validate: weekdays =>
+      isArray(weekdays) ? weekdays : [parseInt(weekdays, 10)],
     test: (day, weekdays) => weekdays.includes(day.weekday),
   },
   ordinalWeekdays: {
     validate: ordinalWeekdays =>
-      Object.keys(ordinalWeekdays)
-        .reduce((obj, ck) => {
-          const weekdays = ordinalWeekdays[ck];
-          if (!weekdays) return obj;
-          obj[ck] = isArray(weekdays) ? weekdays : [parseInt(weekdays, 10)];
-          return obj;
-        }, {}),
+      Object.keys(ordinalWeekdays).reduce((obj, ck) => {
+        const weekdays = ordinalWeekdays[ck];
+        if (!weekdays) return obj;
+        obj[ck] = isArray(weekdays) ? weekdays : [parseInt(weekdays, 10)];
+        return obj;
+      }, {}),
     test: (day, ordinalWeekdays) =>
       Object.keys(ordinalWeekdays)
         .map(k => parseInt(k, 10))
-        .find(k =>
-          ordinalWeekdays[k].includes(day.weekday) &&
-          (k === day.weekdayOrdinal || k === -day.weekdayOrdinalFromEnd)),
+        .find(
+          k =>
+            ordinalWeekdays[k].includes(day.weekday) &&
+            (k === day.weekdayOrdinal || k === -day.weekdayOrdinalFromEnd),
+        ),
   },
   weekends: {
     validate: config => config,
@@ -197,7 +217,8 @@ const _patterns = {
   },
   weeks: {
     validate: weeks => (isArray(weeks) ? weeks : [parseInt(weeks, 10)]),
-    test: (day, weeks) => weeks.includes(day.week) || weeks.includes(-day.weekFromEnd),
+    test: (day, weeks) =>
+      weeks.includes(day.week) || weeks.includes(-day.weekFromEnd),
   },
   months: {
     validate: months => (isArray(months) ? months : [parseInt(months, 10)]),
@@ -208,11 +229,16 @@ const _patterns = {
     test: (day, years) => years.includes(day.year),
   },
 };
-const _patternProps = Object.keys(_patterns).map(k => ({ name: k, validate: _patterns[k].validate }));
+const _patternProps = Object.keys(_patterns).map(k => ({
+  name: k,
+  validate: _patterns[k].validate,
+}));
 const testConfig = (config, day, info) => {
   if (isFunction(config)) return config(day);
   if (isObject(config)) {
-    return Object.keys(config).every(k => _patterns[k].test(day, config[k], info));
+    return Object.keys(config).every(k =>
+      _patterns[k].test(day, config[k], info),
+    );
   }
   return null;
 };
@@ -229,7 +255,9 @@ const DateInfo = (config, order) => {
   if (info.isDate) {
     info.type = 'date';
     // Initialize date from config
-    const date = new Date(config);
+    const date =
+      (!isString(config) && new Date(config)) ||
+      parse(config, defaults.formats.data || ['L', 'YYYY-MM-DD', 'YYYY/MM/DD']);
     // Can't accept invalid dates
     if (isNaN(date)) return null;
     // Strip date time
@@ -245,7 +273,7 @@ const DateInfo = (config, order) => {
     // Date config is a function
     if (isFunction(config)) {
       info.on = { and: config };
-    // Date config is an object
+      // Date config is an object
     } else {
       // Initialize start and end dates (null means infinity)
       let start = config.start && new Date(config.start);
@@ -288,9 +316,8 @@ const DateInfo = (config, order) => {
       }
       // Assign 'or' conditions
       if (config.on) {
-        const or =
-          (isArray(config.on) ? config.on : [config.on])
-          .map((o) => {
+        const or = (isArray(config.on) ? config.on : [config.on])
+          .map(o => {
             if (isFunction(o)) return o;
             const opt = mixinOptionalProps(o, {}, _patternProps);
             return opt.assigned ? opt.target : null;
@@ -305,15 +332,16 @@ const DateInfo = (config, order) => {
   // ========================================================
   // Determines if this date partially intersects another date
   // This is a shallow test (no patterns tested)
-  info.shallowIntersectsDate = other => dateShallowIntersectsDate(info, other.isDate ? other : DateInfo(other));
+  info.shallowIntersectsDate = other =>
+    dateShallowIntersectsDate(info, other.isDate ? other : DateInfo(other));
   // ========================================================
-  info.intersectsDate = (other) => {
+  info.intersectsDate = other => {
     const date = other.isDateInfo ? other : DateInfo(other);
     if (!info.shallowIntersectsDate(date)) return null;
     if (!info.on) return info;
     const range = findShallowIntersectingRange(info, date);
     let result = false;
-    iterateDatesInRange(range, (state) => {
+    iterateDatesInRange(range, state => {
       if (info.matchesDay(state.day)) {
         result = result || date.matchesDay(state.day);
         state.finished = result;
@@ -324,15 +352,16 @@ const DateInfo = (config, order) => {
   // ========================================================
   // Determines if this date completely includes another date
   // This is a shallow test (no patterns tested)
-  info.shallowIncludesDate = other => dateShallowIncludesDate(info, other.isDate ? other : DateInfo(other));
+  info.shallowIncludesDate = other =>
+    dateShallowIncludesDate(info, other.isDate ? other : DateInfo(other));
   // ========================================================
-  info.includesDate = (other) => {
+  info.includesDate = other => {
     const date = other.isDateInfo ? other : DateInfo(other);
     if (!info.shallowIncludesDate(date)) return false;
     if (!info.on) return true;
     const range = findShallowIntersectingRange(info, date);
     let result = true;
-    iterateDatesInRange(range, (state) => {
+    iterateDatesInRange(range, state => {
       if (info.matchesDay(state.day)) {
         result = result && date.matchesDay(state.day);
         state.finished = !result;
@@ -341,7 +370,7 @@ const DateInfo = (config, order) => {
     return result;
   };
   // ========================================================
-  info.includesDay = (day) => {
+  info.includesDay = day => {
     const date = DateInfo(day.date);
     // Date is outside general range - return null
     if (!info.shallowIncludesDate(date)) return null;
@@ -349,13 +378,14 @@ const DateInfo = (config, order) => {
     return info.matchesDay(day) ? info : null;
   };
   // ========================================================
-  info.matchesDay = (day) => {
+  info.matchesDay = day => {
     // No patterns to test
     if (!info.on) return true;
     // Fail if 'and' condition fails
     if (info.on.and && !testConfig(info.on.and, day, info)) return false;
     // Fail if every 'or' condition fails
-    if (info.on.or && !info.on.or.find(or => testConfig(or, day, info))) return false;
+    if (info.on.or && !info.on.or.find(or => testConfig(or, day, info)))
+      return false;
     // Patterns match
     return true;
   };
@@ -372,7 +402,7 @@ const DateInfo = (config, order) => {
     });
   };
   // Build the 'compare to other' function
-  info.compare = (other) => {
+  info.compare = other => {
     if (info.order !== other.order) return info.order - other.order;
     if (info.type !== other.type) return info.isDate ? 1 : -1;
     if (info.isDate) return 0;