+++ test/unit/specs/Calendar.spec.js
@@ -1,16 +1,16 @@
-import { shallow } from '@vue/test-utils';
-import Calendar from '@/components/Calendar';
+// import { shallow } from '@vue/test-utils';
+// import Calendar from '@/components/Calendar';
 
 describe('Calendar', () => {
   it('renders props.fromPage', () => {
-    const fromPage = { month: 4, year: 2001 };
-    const wrapper = shallow(Calendar, {
-      propsData: {
-        fromPage,
-      },
-    });
-    const cal = wrapper.find('div');
-    console.log(cal.html());
+    // const fromPage = { month: 4, year: 2001 };
+    // const wrapper = shallow(Calendar, {
+    //   propsData: {
+    //     fromPage,
+    //   },
+    // });
+    // const cal = wrapper.find('div');
+    // console.log(cal.html());
     // const input = cal.find('input');
     // console.log(input);
     // if (cal) {