+++ test/unit/specs/DatePicker.spec.js
@@ -1,17 +1,17 @@
-import { mount } from '@vue/test-utils';
-import DatePicker from '@/components/DatePicker';
+// import { mount } from '@vue/test-utils';
+// import DatePicker from '@/components/DatePicker';
 
 describe('DatePicker', () => {
   it('renders props.fromPage', () => {
-    const fromPage = { month: 4, year: 2001 };
-    const wrapper = mount(DatePicker, {
-      propsData: {
-        fromPage,
-        isInline: true,
-      },
-    });
-    const cal = wrapper.find('div');
-    console.log(cal.html());
-    expect(2).toBe(2);
+    // const fromPage = { month: 4, year: 2001 };
+    // const wrapper = mount(DatePicker, {
+    //   propsData: {
+    //     fromPage,
+    //     isInline: true,
+    //   },
+    // });
+    // const cal = wrapper.find('div');
+    // console.log(cal.html());
+    // expect(2).toBe(2);
   });
 });